requirements.txt
# requirements.txt
setuptools==68.2.2
py2app
PySide6
Pillow